edit: Include user profile in user credentials functions

+++ app/controllers/user_credentials/user_credentials_controller.js
@@ -10,13 +10,24 @@ const UserCredentialsController = {
   create: async (req, res) => {
     await sequelize.transaction(async(t) => {
       try {
-        const user_credentials = await UserCredentials.create({
-          'profile_id': req.body.profile_id,
-          'username': req.body.username,
-          'password': req.body.password,
-          'email': req.body.email,
-          'phone': req.body.phone,
-        });
+        const user_profile = {
+          'first_name': req.body.first_name,
+          'middle_name': req.body.middle_name,
+          'last_name': req.body.last_name,
+          'department_id': req.body.department_id,
+          'role_id': req.body.role_id,
+          'profile_photo': req.body.profile_photo,
+        }
+        const user_credentials = await UserCredentials.create(
+          {
+            'username': req.body.username,
+            'password': req.body.password,
+            'email': req.body.email,
+            'phone': req.body.phone,
+            'UserProfile': user_profile,
+          },
+          { include: [UserProfile] }, { transaction: t }
+        );
         return res.status(CREATED).json({UserCredentials: user_credentials})
       } catch(error) {
         return res.status(INTERNAL_SERVER_ERROR).json({ message: error.message })
@@ -35,11 +46,11 @@ const UserCredentialsController = {
               include: [
                 {
                   model: Department,
-                  attributes: ['department_name']
+                  attributes: ['id','department_name']
                 },
                 {
                   model: Role,
-                  attributes: ['role_name']
+                  attributes: ['id','role_name']
                 }
               ]
             }
@@ -54,10 +65,13 @@ const UserCredentialsController = {
             password: account.password,
             email: account.email,
             phone: account.phone,
+            profile_id: profile?.id,
             first_name: profile?.first_name,
             middle_name: profile?.middle_name,
             last_name: profile?.last_name,
+            department_id: profile?.Department?.id,
             department_name: profile?.Department?.department_name,
+            role_id: profile?.Role?.id,
             role_name: profile?.Role?.role_name,
             profile_photo: profile?.profile_photo,
           }
@@ -129,17 +143,29 @@ const UserCredentialsController = {
           where: { id: req.params.id }
         });
 
-        if (!user_credentials) {
-          return res.status(NOT_FOUND).json({ message: 'User Credentials Not Found' });
+        const user_profile = await UserProfile.findOne({
+          where: { id: req.body.profile_id }
+        });
+
+        if (!user_credentials || !user_profile) {
+          return res.status(NOT_FOUND).json({ message: 'User Not Found' });
         }
 
         await user_credentials.update({
-          profile_id: req.body.profile_id,
           username: req.body.username,
           password: req.body.password,
           email: req.body.email,
           phone: req.body.phone,
-        }, {transaction: t})
+        }, { transaction: t })
+
+        await user_profile.update({
+          first_name: req.body.first_name,
+          middle_name: req.body.middle_name,
+          last_name: req.body.last_name,
+          department_id: req.body.department_id,
+          role_id: req.body.role_id,
+          profile_photo: req.body.profile_photo,
+        }, { transaction: t })
 
         return res.status(OK).json({UserCredentials: user_credentials});
       } catch(error) {
@@ -154,14 +180,22 @@ const UserCredentialsController = {
           where: { id: req.params.id }
         })
 
-        if (!user_credentials) {
-          return res.status(NOT_FOUND).json({ message: 'User Credentials Not Found' })
+        const user_profile = await UserProfile.findOne({
+          where: { id: user_credentials.profile_id }
+        })
+
+        if (!user_credentials || !user_profile) {
+          return res.status(NOT_FOUND).json({ message: 'User Not Found' })
         }
 
         await user_credentials.destroy({
           force: false
         })
-        return res.status(OK).json({ message: 'User Credentials Destroyed' })
+
+        await user_profile.destroy({
+          force: false
+        })
+        return res.status(OK).json({ message: 'User Destroyed' })
       } catch(error) {
         return res.status(INTERNAL_SERVER_ERROR).json({ message: error.message })
       }
